backend/requirements.txt
fastapi
uvicorn
boto3
google-auth
google-api-python-client
azure-identity
azure-mgmt-costmanagement
azure-mgmt-resource
python-dotenv
sqlalchemy
psycopg2-binary
alembic
pydantic-settings
redis
celery
httpx
requests